1.0.1 BETA
Summary:
Fixed: Config auto-test that does not check the correct folders
New: kernel_shutdown to allow modules to trigger the normal framework
shutdown event

Full description of this update:

Config auto-test: The config auto-test was buggy. lt does check if
folders exists but in the wrong directory. It is now checking the
app_real_location first and the other folder from that directory. In the
previous version, if the app_real_location’s folder exists (whatever it
is) the
framework try to use it, but give an error then crash (kernel panic).

Kernel: It is now possible to trigger the framework shutdown without
using exit(). That allow modules to stop execution and save the log if
debug is enabled. It also trigger the functions in the shutdown array
(allowing other modules to work normally). The end of the index.php file
was moved to a new function called kerneL shutdown in the kernel’s basic
system (boot.php).