1.0.2 BETA
Summary:
Fixed: sql_query always return a 1 row array

Details of this update:
This is an urgent update. When executing a query that use SELECT, an
array is returned to the caller but it always contain the first row and
never use the remaining one. This is caused by a missing array_push
directive before returning the value.