1.1 BETA
Summary:
Major release. This update improve the framework a lot and contain some
bugfixes. It also introduce the first web-based configuration creator
for the framework.

IMPORTANT: Since we are now calling a function to render the page, the
$THEME variable is no longuer directly accessible from the theme. The
default theme has been updated to use the $GLOBALS variable for getting
the page title

Details of this update:

Added: setup.php. This file is the configuration wizard. No more need to
configure each options manually. For now, it use a really basic
interface and allow you to modify any setting. Further version will have
a better look and less options (remove those that can be detected
dynamically)

Fix: Small bug when running the framework at the root of the web server
Fix: 404 bug when running the framework at the root of the web server
(not the same fix as the above line)

Modification: One word changed in the kernel panic message

Fix: Files download was not working properly

Fix: Log file (debug.log) is now live. In case of an unexpected
execution stop, the log was not saved. This was making debugging a lot
harder. The log is now written each time a log entry is added

Added: PHP errors are now saved to kernel_log and shown with an "An
error occured" error message in the web page if debug is enabled.

Improvement: Page rendering has been rewritten. It is now possible for
modules to trigger the normal page rendering or file download (depends
on the request) by calling __render_page (function name may be modified
in a few updates). This also fixed a lots of problems when trying to
download files (for some reason, page rendering would not work and no
file were sent to the browser [blank page])

Improvement: It can be hard to find the moment that the framework boot
in the log. There is a now a space at the last line when calling
kernel_shutdown

Fix: app_location is now tested at each start. This make sure we are
where we are supposed to be.