HOT FIX: Fix bug when accessing pages
!This is a hot fix!
The last version had a bug when trying to use any pages. When accessing the framework root, all is good but when trying to access a specific page, we had to add a double slash before the page's name.
This bug should have been detected before the release of the latest version but it wasn't. Sorry for any inconvenience this may have caused.