1.3 BETA
Added MODULESLOADED event. This allow the use of functions included from other modules
Added kernel_checkIfModuleIsValid. This allow a module to check if another module exist and is valid. Can be used to check for dependency
Added kernel_checkIfModuleExist. This is an alias of kernel_checkIfModuleIsValid
Theme's files are now processed as part of the framework:
   This allow pre-processing of all files including JS and CSS.
   The REAL theme folder is hidden from the public. This prevent users from accessing other theme's files and a static public path (can be configured in config.php)
   Basic functions added to default theme
MySQLi driver updated. SQL ressource ID is now saved when using sql_connect
Copyright year updated to 2015