2.0 RELEASE
We're now OUT OF BETA!

It took us (or me... actually) so long before releasing this new version (I won't explain why here), but this was for the best.
This 2.0 release is now the first stable one and all bugs should now be fixed. Feel free to report any new bugs or request features via the Issue tracker!

Here's a summary of the changes within this version:
- We now have a Wiki! It contains a complete documentation for all the functions and a "Getting Started" guide (more guides are planned later)
- A more realistic "default" theme is now included
- Some example pages are now provided within the "webroot" folder (to be used as a part of the "Getting Started" guide)
- Many bug fixes and improvements (both to code readability and performance)
- Various typos and errors in strings corrected
- Better setup.php (more details for the debugging section)